pyproject.toml
[project]
name = "tmint-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "python-telegram-bot>=20.7",
  "notion-client",
  "python-dotenv",
  "groq>=0.4.0",
   "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
]
requires = ["hatchling"]
build-backend = "hatchling.build"
